+++ gen/com/kaylerrenslow/a3plugin/lang/header/parser/HeaderParser.java
@@ -396,14 +396,12 @@ public class HeaderParser implements PsiParser, LightPsiParser {
   }
 
   /* ********************************************************** */
-  // INLINE_COMMENT | BLOCK_COMMENT | preprocessor_group | statement_
+  // preprocessor_group | statement_
   public static boolean file_entry(PsiBuilder b, int l) {
     if (!recursion_guard_(b, l, "file_entry")) return false;
     boolean r;
     Marker m = enter_section_(b, l, _NONE_, FILE_ENTRY, "<file entry>");
-    r = consumeToken(b, INLINE_COMMENT);
-    if (!r) r = consumeToken(b, BLOCK_COMMENT);
-    if (!r) r = preprocessor_group(b, l + 1);
+    r = preprocessor_group(b, l + 1);
     if (!r) r = statement_(b, l + 1);
     exit_section_(b, l, m, r, false, null);
     return r;